+++ rusk-recovery/Makefile
@@ -2,10 +2,10 @@ help: ## Display this help screen
 	@grep -h -E '^[a-zA-Z_-]+:.*?## .*$$' $(MAKEFILE_LIST) | awk 'BEGIN {FS = ":.*?## "}; {printf "\033[36m%-15s\033[0m %s\n", $$1, $$2}'
 
 keys: ## Build circuit keys
-	cargo run --release -- keys
+	cargo run --release --bin rusk-recovery-keys --features keys
 
 state: ## Build network state
-	cargo run --release -- state
+	cargo run --release --bin rusk-recovery-state --features state
 
 test: ## Run Rusk tests
 	@cargo test \
